BUG: Prevent segfault on cleanup
In case an object (i.e. QTimer) is created with a parent like QApplication,
and considering the fact that PythonQt::cleanup() is called before
qSlicerApplication is completely destroyed, the application will crashes.

Indeed, the associated PythonQtShell object (ie PythonQtShell_QTimer)
is parented with QApplication. This object will then be deleted when the
program exit and after PythonQt has been cleanup. Internally the
PythonQtShell objetct refers to PythonQt::priv() and also expect
Python to be in a "initialized" state.

This commit had checks to make sure PythonQtPrivate is valid and also to verify
that Python is still initialized before calling function like Py_DECREF.
Note that these checks have been added just in the location that were causing problem.

An other approach could be to transform PythonQt into a real singleton
( for example by inspiring from ctkSingleton helper that implements
the nifty counter idiom). Nevertheless, such approach will require to
refactor the way PythonQt is initialized.