[Backport] Make __enter__ and __exit__ slots work with "with" statements
The problem was that the "with" statement doesn't use the conventional __getattr__ access, but directly accesses the tp_dict of the type, which usually is empty in PythonQt, so we had to put the __enter__ and __exit__ methods there, too.
And for binding the methods to the instance, tp_descr_get was used in this case, so we had to implement that for slots, too.

Discussed with and reviewed by Florian Link

(cherry picked from commit MeVisLab/pythonqt@d13b66c8b8c82f4aaf2bab6b11107db9f45b63c8)