[Backport] Fix crash when using QObject::disconnect method (when multi-threading is enabled)
It seems this is done with a slot call, which gives up the GIL, but the GIL is needed in the destructor of the PythonQtSignalReceiver when a callable target was given.
Fix this by simply using a PythonQtSafeObjectPtr for the _callable.

(cherry picked from commit MeVisLab/pythonqt@5fc0dfc52c554d734a3cf306e05cdfa3a97f2098)