[Backport generator] Handle enum classes
Since the enum values can be duplicates (they only need to be unique in the scope of the enum class) I opted to prefix them with the enum class name (and an underscore) in the generated wrapper.
It would have been nice to use a Python Enum class instead, but after a quick look this seemed to be quite a lot of work.

(cherry picked from commit MeVisLab/pythonqt@c707bcceb5617726e25ff8119fe86817cc7bda53)